Move continuation detection into eval
With this change, eval can return the output of running the code, or
a sentinel value that indicates that more input is needed to run a chunk
of code.  This is so that more advanced forms of detecting when we need
more input, such as if the user is entering input for a multi-line
statement like a block.

This is just a proof-of-concept change; I don't really like how eval
indicates to its caller that it needs more input.  But at least this
will get people to try and see if multi-line input could be feasible.

We also add a new method to HLL::Compiler - needs-more-input, which is
used in conjunction with the sentinel value returned by eval.